Feature - Loading JSON file
Created code to open and read the information on the JSON file and store
it as variables. I also created getters and setters for values that can
change and also created a file in testing to see if the JSON file loads
or not. There are also some test JSON files that are added.