Feature: Creating JSON Files
Created the ability to generate information on the user and store it as
a JSON file. I also made sure to check if the username would be valid by
creating a separate method to check the username. There are also code to
see how it works and to test the code.