Feature: Save user data to JSON file
Created code to be able to save information of the user to the JSON file
when the game is being played. Also refactored creating the JSON so when
viewing the JSON file, it is easier to read. Created two new JSON files
as we are required to have four users and also created testing for
saving the JSON file