Fix: UI lag on first drawing of session mitigated.
- Added a "ghost game" which is run when the canvas scene is loaded from
the GameSelectController. This invisible game is loaded, run, and reset
behing the loading screen to initialise the game services which
otherwise cause a small blip of lag on the first drawing of a session.
- Refactored text to speech task into service which can be cancelled and
reset. This allows the text-to-speech service to also be pre-loaded
alongside the game service.
- Added isGhostGame flag to Game class, to avoid accidental wins or
losses for the player while running the ghost game.